Finalização do Mensalista

+++ CrudePOO/src/pkg1/Valor.java
@@ -1,5 +1,7 @@
 package pkg1;
 
+import java.math.BigDecimal;
+
 public interface Valor {
-	double calcularValor();
+	BigDecimal calcularValor();
 }
